api/requirements.txt
appdirs==1.4.3
appnope==0.1.0
argh==0.26.2
colorama==0.3.9
decorator==4.0.11
Django==1.11.1
django-extensions==1.7.9
django-fixture-magic==0.1.3
djangorestframework==3.6.2
djangorestframework-jsonapi==2.2.0
djangorestframework-jwt==1.10.0
docopt==0.6.2
drf-nested-routers==0.90.0
inflection==0.3.1
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
packaging==16.8
pathtools==0.1.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.33
Pygments==2.2.0
PyJWT==1.5.0
pyobjc-core==3.2.1
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
pytest-osxnotify==0.1.7
pytest-watch==4.1.0
pytz==2017.2
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
watchdog==0.8.3
wcwidth==0.1.7
